requirements.txt
numpy
scipy
pandas
h5py
numpy-quaternion
matplotlib
torch
torchvision
tensorboardX
numba
plyfile
tqdm
scikit-learn
